Added Timer
I added a timer that pauses and unpauses when the user enters or leaves
the game. I made a screen popup when the user is out of time and a reset
button that resets the timer and wires. I also changed all the wires to
white and then used Corona's setFillColor function to color them instead
of having the colors be fixed by the png image